Implemented Exponential Backoff

+++ CSULB.GetUsGrub/CSULB.GetUsGrub.Models/Models/Interfaces/IGeoCoordinates.cs
@@ -2,7 +2,7 @@
 {
     public interface IGeoCoordinates
     {
-        float Latitude { get; set; }
-        float Longitude { get; set; }
+        double Latitude { get; set; }
+        double Longitude { get; set; }
     }
 }
